+++ src/main/java/fr/elecmove/api/controller/dto/location_station/LocationStationDTO.java
@@ -1,13 +1,15 @@
 package fr.elecmove.api.controller.dto.location_station;
 
+import java.math.BigDecimal;
+
 public class LocationStationDTO {
 
     private String id;
     private String address;
     private String city;
     private String zipcode;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
 
     public String getId() {
@@ -42,19 +44,19 @@ public class LocationStationDTO {
         this.zipcode = zipcode;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
 }
